refactor: Move runCustomBuild (and associated functions) to its own file
