+++ packages/wrangler/src/deployment-bundle/entry.ts
@@ -1,11 +1,10 @@
 import assert from "node:assert";
-import { existsSync, statSync } from "node:fs";
 import path from "node:path";
 import * as esbuild from "esbuild";
-import { execaCommand } from "execa";
 import { logger } from "../logger";
 import { getBasePath } from "../paths";
 import { COMMON_ESBUILD_OPTIONS } from "./bundle";
+import { runCustomBuild } from "./run-custom-build";
 import type { Config } from "../config";
 import type { DurableObjectBindings } from "../config/environment";
 import type { CfScriptFormat } from "./worker";
@@ -72,15 +71,6 @@ export async function getEntry(
 	const relativeFile = path.relative(directory, file) || ".";
 	await runCustomBuild(file, relativeFile, config.build);
 
-	if (fileExists(file) === false) {
-		throw new Error(
-			getMissingEntryPointMessage(
-				`The entry-point file at "${relativeFile}" was not found.`,
-				file,
-				relativeFile
-			)
-		);
-	}
 	const format = await guessWorkerFormat(
 		file,
 		directory,
@@ -122,36 +112,6 @@ export async function getEntry(
 		moduleRoot: args.moduleRoot ?? config.base_dir ?? path.dirname(file),
 		name: config.name ?? "worker",
 	};
-}
-
-export async function runCustomBuild(
-	expectedEntryAbsolute: string,
-	expectedEntryRelative: string,
-	build: Config["build"]
-) {
-	if (build?.command) {
-		// TODO: add a deprecation message here?
-		logger.log("Running custom build:", build.command);
-		await execaCommand(build.command, {
-			shell: true,
-			// we keep these two as "inherit" so that
-			// logs are still visible.
-			stdout: "inherit",
-			stderr: "inherit",
-			...(build.cwd && { cwd: build.cwd }),
-		});
-
-		if (fileExists(expectedEntryAbsolute) === false) {
-			throw new Error(
-				getMissingEntryPointMessage(
-					`The expected output file at "${expectedEntryRelative}" was not found after running custom build: ${build.command}.\n` +
-						"The `main` property in wrangler.toml should point to the file generated by the custom build.",
-					expectedEntryAbsolute,
-					expectedEntryRelative
-				)
-			);
-		}
-	}
 }
 
 /**
@@ -237,23 +197,6 @@ function listEntryPoints(
 type ValueOf<T> = T[keyof T];
 
 /**
- * Returns true if the given `filePath` exists as-is,
- * or if some version of it (by appending a common extension) exists.
- */
-function fileExists(filePath: string): boolean {
-	if (path.extname(filePath) !== "") {
-		return existsSync(filePath);
-	}
-	const base = path.join(path.dirname(filePath), path.basename(filePath));
-	for (const ext of [".ts", ".tsx", ".js", ".jsx"]) {
-		if (existsSync(base + ext)) {
-			return true;
-		}
-	}
-	return false;
-}
-
-/**
  * Groups the durable object bindings into two lists:
  * those that are defined locally and those that refer to a durable object defined in another script.
  */
@@ -294,51 +237,4 @@ function generateAddScriptNameExamples(
 			return `${currentBinding} ==> ${fixedBinding}`;
 		})
 		.join("\n");
-}
-
-/**
- * Generate an appropriate message for when the entry-point is missing.
- *
- * To be more helpful to developers, we check whether there is a suitable file
- * nearby to the expected file path.
- */
-function getMissingEntryPointMessage(
-	message: string,
-	absoluteEntryPointPath: string,
-	relativeEntryPointPath: string
-): string {
-	if (
-		existsSync(absoluteEntryPointPath) &&
-		statSync(absoluteEntryPointPath).isDirectory()
-	) {
-		// The expected entry-point is a directory, so offer further guidance.
-		message += `\nThe provided entry-point path, "${relativeEntryPointPath}", points to a directory, rather than a file.\n`;
-
-		// Perhaps we can even guess what the correct path should be...
-		const possiblePaths: string[] = [];
-		for (const basenamePath of [
-			"worker",
-			"dist/worker",
-			"index",
-			"dist/index",
-		]) {
-			for (const extension of [".ts", ".tsx", ".js", ".jsx"]) {
-				const filePath = basenamePath + extension;
-				if (fileExists(path.resolve(absoluteEntryPointPath, filePath))) {
-					possiblePaths.push(path.join(relativeEntryPointPath, filePath));
-				}
-			}
-		}
-
-		if (possiblePaths.length > 0) {
-			message +=
-				`\nDid you mean to set the main field to${
-					possiblePaths.length > 1 ? " one of" : ""
-				}:\n` +
-				"```\n" +
-				possiblePaths.map((filePath) => `main = "./${filePath}"\n`).join("") +
-				"```";
-		}
-	}
-	return message;
 }
